backend/requirements.txt
Django==6.0.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg[binary]
requests==2.32.3
gunicorn==22.0.0